+++ src/main/java/frc/robot/subsystems/ConditionedDriveSubsystem.java
@@ -1,15 +1,14 @@
 package frc.robot.subsystems;
 
-import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import org.usfirst.frc.team4999.utils.Utils;
 
-public class ConditionedDriveSubsystem extends SubsystemBase {
+public class ConditionedDriveSubsystem extends DriveSubsystem {
   private double spd_limit;
   private final double[] spd_limits = { 0.25, 0.33333, 0.5, 0.66666, 0.75, 1 };
   private int spd_limit_index;
-  private DriveSubsystem m_subsystem;
+  private FalconDriveSubsystem m_subsystem;
 
-  public ConditionedDriveSubsystem(DriveSubsystem subsystem) {
+  public ConditionedDriveSubsystem(FalconDriveSubsystem subsystem) {
     spd_limit_index = spd_limits.length - 1;
     m_subsystem = subsystem;
   }
@@ -18,7 +17,7 @@ public class ConditionedDriveSubsystem extends SubsystemBase {
     spd_limit = limit;
   }
 
-  public void mapToLimit(double moveRequest, double turnRequest) {
+  public void drive(double moveRequest, double turnRequest) {
     m_subsystem.drive(Utils.map(moveRequest, -1, 1, -spd_limit, spd_limit), turnRequest);
   }
 
@@ -34,5 +33,9 @@ public class ConditionedDriveSubsystem extends SubsystemBase {
     if (spd_limit_index > 0)
       spd_limit_index = 0;
     spd_limit = spd_limits[spd_limit_index];
+  }
+
+  public void stop() {
+
   }
 }
